+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Changes:
 
 - Switched builds from CircleCI to Github Actions ([273](https://github.com/kpumuk/meta-tags/pull/273))
 
+Changes:
+
+- Ruby on Rails < 6.0 is no longer supported.
+
 ## 2.19.0 (October 5, 2023) [☰](https://github.com/kpumuk/meta-tags/compare/v2.18.0...v2.19.0)
 
 Changes:
@@ -190,7 +194,7 @@ Features:
 
 Changes:
 
-- Rails < 3.2 is not longer supported
+- Ruby on Rails < 3.2 is no longer supported
 
 Features:
 
@@ -209,7 +213,7 @@ Bugfixes:
 
 Changes:
 
-- Ruby < 2.0 is not longer supported
+- Ruby < 2.0 is no longer supported
 
 Features: